# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "NUS API Documentation",
        "description": "NUS Swagger OpenApi description",
        "contact": {
            "email": "ryanhaidypapudi@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/topics": {
            "get": {
                "tags": [
                    "Topic"
                ],
                "summary": "Get list of topics",
                "description": "",
                "operationId": "GetTopics",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Topic"
                ],
                "summary": "Store topic to system",
                "description": "",
                "operationId": "PostTopic",
                "requestBody": {
                    "description": "Topic store request object",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/TopicStoreRequest"
                            }
                        },
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TopicStoreRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/topics/{slug}": {
            "get": {
                "tags": [
                    "Topic"
                ],
                "summary": "Get topic by slug",
                "description": "",
                "operationId": "GetTopic",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "Slug of topic to return",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Topic"
                ],
                "summary": "Update topic in system",
                "description": "",
                "operationId": "PutTopic",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "Slug of topic to update",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Topic update request object",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/TopicUpdateRequest"
                            }
                        },
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TopicUpdateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Topic"
                ],
                "summary": "Delete topic in system",
                "description": "",
                "operationId": "DeleteTopic",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "Slug of topic to update",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/topics/join": {
            "post": {
                "tags": [
                    "Topic"
                ],
                "summary": "Add/Remove user's joined topic in system",
                "description": "",
                "operationId": "JoinTopic",
                "requestBody": {
                    "description": "Topic join request object",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/TopicJoinRequest"
                            }
                        },
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TopicJoinRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "TopicJoinRequest": {
                "properties": {
                    "topic_id": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "TopicStoreRequest": {
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "background": {
                        "type": "file"
                    }
                },
                "type": "object"
            },
            "TopicUpdateRequest": {
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "background": {
                        "type": "file"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "Bearer"
            }
        }
    },
...
    "security": [
        {
            "bearer": {
                "type": "http",
                "description": "Authorization token obtained from logging in.",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer"
            }
        }
    ]
}